+++ apps/dashboard/src/@/components/ui/tracked-link.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useTrack } from "hooks/analytics/useTrack";
 import Link from "next/link";
 import type React from "react";
