+++ apps/dashboard/src/components/embedded-wallets/index.tsx
@@ -1,27 +1,32 @@
+"use client";
+
 import { TabButtons } from "@/components/ui/tabs";
 import type { ApiKey } from "@3rdweb-sdk/react/hooks/useApi";
-import type { EmbeddedWalletUser } from "@3rdweb-sdk/react/hooks/useEmbeddedWallets";
+import { useSearchParams } from "next/navigation";
 import { useState } from "react";
 import { Configure } from "./Configure";
 import { Users } from "./Users";
 
 interface EmbeddedWalletsProps {
-  apiKey: ApiKey;
-  wallets: EmbeddedWalletUser[];
-  isLoading: boolean;
-  isFetched: boolean;
+  apiKey: Pick<
+    ApiKey,
+    | "id"
+    | "name"
+    | "domains"
+    | "bundleIds"
+    | "services"
+    | "redirectUrls"
+    | "key"
+  >;
   trackingCategory: string;
-  defaultTabIndex?: number;
 }
 
 export const EmbeddedWallets: React.FC<EmbeddedWalletsProps> = ({
   apiKey,
-  wallets,
-  isLoading,
-  isFetched,
   trackingCategory,
-  defaultTabIndex,
 }) => {
+  const searchParams = useSearchParams();
+  const defaultTabIndex = searchParams?.get("tab") === "1" ? 1 : 0;
   const [selectedTab, setSelectedTab] = useState<"users" | "config">(
     defaultTabIndex === 0 ? "users" : "config",
   );
@@ -60,12 +65,7 @@ export const EmbeddedWallets: React.FC<EmbeddedWalletsProps> = ({
       <div className="h-6" />
 
       {selectedTab === "users" && (
-        <Users
-          wallets={wallets}
-          isLoading={isLoading}
-          isFetched={isFetched}
-          trackingCategory={trackingCategory}
-        />
+        <Users clientId={apiKey.key} trackingCategory={trackingCategory} />
       )}
 
       {selectedTab === "config" && (
