+++ apps/dashboard/src/pages/dashboard/connect/in-app-wallets.tsx
@@ -1,26 +1,23 @@
 import { Spinner } from "@/components/ui/Spinner/Spinner";
 import { TrackedLinkTW } from "@/components/ui/tracked-link";
 import { type ApiKey, useApiKeys } from "@3rdweb-sdk/react/hooks/useApi";
-import { useEmbeddedWallets } from "@3rdweb-sdk/react/hooks/useEmbeddedWallets";
 import { useLoggedInUser } from "@3rdweb-sdk/react/hooks/useLoggedInUser";
 import { AppLayout } from "components/app-layouts/app";
 import { EmbeddedWallets } from "components/embedded-wallets";
 import { ApiKeysMenu } from "components/settings/ApiKeys/Menu";
 import { NoApiKeys } from "components/settings/ApiKeys/NoApiKeys";
 import { ConnectSidebar } from "core-ui/sidebar/connect";
-import { ArrowRightIcon } from "lucide-react";
 import { useRouter } from "next/router";
 import { PageId } from "page-id";
 import { useMemo, useState } from "react";
 import type { ThirdwebNextPage } from "utils/types";
-import { Analytics } from "../../../components/embedded-wallets/Users/Analytics";
-import { SupportedPlatformLink } from "../../../components/wallets/SupportedPlatformLink";
+import { AnalyticsCallout } from "../../../app/team/[team_slug]/[project_slug]/connect/in-app-wallets/_components/AnalyticsCallout";
+import { InAppWaletFooterSection } from "../../../app/team/[team_slug]/[project_slug]/connect/in-app-wallets/_components/footer";
 
 const TRACKING_CATEGORY = "embedded-wallet";
 
 const DashboardConnectEmbeddedWallets: ThirdwebNextPage = () => {
   const router = useRouter();
-  const defaultTabIndex = Number.parseInt(router.query.tab?.toString() || "0");
   const defaultClientId = router.query.clientId?.toString();
   const { isLoading } = useLoggedInUser();
   const keysQuery = useApiKeys();
@@ -50,10 +47,6 @@ const DashboardConnectEmbeddedWallets: ThirdwebNextPage = () => {
     }
     return undefined;
   }, [apiKeys, defaultClientId, selectedKey_]);
-
-  const walletsQuery = useEmbeddedWallets(selectedKey?.key as string);
-
-  const wallets = walletsQuery?.data || [];
 
   if (isLoading) {
     return (
@@ -113,174 +106,20 @@ const DashboardConnectEmbeddedWallets: ThirdwebNextPage = () => {
           {hasApiKeys && selectedKey && (
             <EmbeddedWallets
               apiKey={selectedKey}
-              wallets={wallets}
-              isLoading={walletsQuery.isLoading}
-              isFetched={walletsQuery.isFetched}
               trackingCategory={TRACKING_CATEGORY}
-              defaultTabIndex={defaultTabIndex}
             />
           )}
         </>
       )}
 
       <div className="h-16" />
-      <Analytics trackingCategory={TRACKING_CATEGORY} />
+      <AnalyticsCallout trackingCategory={TRACKING_CATEGORY} />
       <div className="h-5" />
 
-      <div className="grid grid-cols-1 md:grid-cols-2 gap-5">
-        <ViewDocs />
-        <Templates />
-      </div>
+      <InAppWaletFooterSection trackingCategory={TRACKING_CATEGORY} />
     </div>
   );
 };
-
-function ViewDocs() {
-  return (
-    <div className="p-4 lg:p-6 bg-muted/50 border border-border rounded-lg">
-      <div className="flex items-center gap-2">
-        <h3 className="font-semibold">View Docs</h3>
-        <ArrowRightIcon className="size-4" />
-      </div>
-
-      <div className="h-4" />
-
-      <div className="grid grid-cols-2 gap-4">
-        <SupportedPlatformLink
-          trackingCategory={TRACKING_CATEGORY}
-          platform="React"
-          href="https://portal.thirdweb.com/connect/in-app-wallet/overview"
-        />
-
-        <SupportedPlatformLink
-          trackingCategory={TRACKING_CATEGORY}
-          platform="Unity"
-          href="https://portal.thirdweb.com/unity/wallets/providers/embedded-wallet"
-        />
-        <SupportedPlatformLink
-          trackingCategory={TRACKING_CATEGORY}
-          platform="React Native"
-          href="https://portal.thirdweb.com/react/v5/in-app-wallet/get-started"
-        />
-        <SupportedPlatformLink
-          trackingCategory={TRACKING_CATEGORY}
-          platform="TypeScript"
-          href="https://portal.thirdweb.com/connect/in-app-wallet/overview"
-        />
-      </div>
-
-      <div className="h-6" />
-
-      <div className="flex items-center gap-2">
-        <h3 className="font-semibold">Relevant Guides</h3>
-        <ArrowRightIcon className="size-4" />
-      </div>
-
-      <div className="h-4" />
-
-      <div className="flex flex-col gap-3">
-        <GuideLink
-          href="https://blog.thirdweb.com/what-are-embedded-wallets/"
-          label="what-is-an-embedded-wallet"
-        >
-          What is an in-app wallet?
-        </GuideLink>
-
-        <GuideLink
-          href="https://portal.thirdweb.com/connect/in-app-wallet/get-started"
-          label="sdks-get-started"
-        >
-          Get started with In-App Wallets
-        </GuideLink>
-
-        <GuideLink
-          href="https://portal.thirdweb.com/connect/in-app-wallet/how-to/connect-users"
-          label="how-to-connect-your-users"
-        >
-          Using In-App Wallets with Connect
-        </GuideLink>
-
-        <GuideLink
-          href="https://portal.thirdweb.com/connect/in-app-wallet/how-to/build-your-own-ui"
-          label="how-to-build-your-own-ui"
-        >
-          How to Build Your Own UI
-        </GuideLink>
-
-        <GuideLink
-          href="https://portal.thirdweb.com/connect/in-app-wallet/custom-auth/custom-auth-server"
-          label="how-to-custom-auth-server"
-        >
-          Create a custom auth server
-        </GuideLink>
-      </div>
-    </div>
-  );
-}
-
-function GuideLink(props: {
-  label: string;
-  children: React.ReactNode;
-  href: string;
-}) {
-  return (
-    <TrackedLinkTW
-      category={TRACKING_CATEGORY}
-      label={"guide"}
-      trackingProps={{
-        guide: props.label,
-      }}
-      href={props.href}
-      className="text-sm !text-muted-foreground hover:!text-foreground"
-      target="_blank"
-    >
-      {props.children}
-    </TrackedLinkTW>
-  );
-}
-
-function Templates() {
-  return (
-    <div className="p-4 lg:p-6 bg-muted/50 border border-border rounded-lg">
-      <div className="flex items-center gap-2">
-        <h3 className="font-semibold">Relevant Templates</h3>
-        <ArrowRightIcon className="size-4" />
-      </div>
-
-      <div className="h-6" />
-
-      <div className="flex flex-col gap-3">
-        <GuideLink
-          href="https://github.com/thirdweb-example/embedded-smart-wallet"
-          label="embedded-smart-wallet"
-        >
-          In-App Wallet + Account Abstraction Starter Kit
-        </GuideLink>
-
-        <GuideLink
-          href="https://github.com/thirdweb-example/catattacknft"
-          label="embedded-cat-attack"
-        >
-          Cat Attack [Demo Web Game]
-        </GuideLink>
-
-        <GuideLink
-          href="https://github.com/thirdweb-example/embedded-wallet-custom-ui"
-          label="embedded-wallet-with-custom-ui-react"
-        >
-          In-App Wallet With Custom UI [React]
-        </GuideLink>
-
-        <GuideLink
-          href="https://github.com/thirdweb-example/embedded-wallet-custom-ui-react-native"
-          label="embedded-wallet-with-custom-ui-react-native"
-        >
-          In-App Wallet With Custom UI [ReactNative]
-        </GuideLink>
-      </div>
-    </div>
-  );
-}
 
 DashboardConnectEmbeddedWallets.getLayout = (page, props) => (
   <AppLayout {...props} hasSidebar={true}>
